Introduce MFEMWeakFormProblemComposerBase to avoid introducing EquationSystem requirements into ProblemOperator base class
