+++ framework/doc/content/source/mfem/problem_composers/MFEMProblemComposer.md
@@ -9,7 +9,10 @@ objects, through the method `createProblemOperator()`. The composer itself is cr
 the [MFEMProblem.md].
 
 These composer objects can be used to instantiate the provided equation system problem operators, or user defined 
-custom ones. For example usage on custom operators refer to the walkthrough example in [this page](syntax/ProblemComposers/index.md).
+custom ones. Composers whose operator solves an equation system built from the
+[`WeakForms`](syntax/WeakForms/index.md) block derive from
+[MFEMWeakFormProblemComposerBase.md] instead, which adds the `weak_form` parameter used to select
+that system. For example usage on custom operators refer to the walkthrough example in [this page](syntax/ProblemComposers/index.md).
 
 !if-end!
