+++ framework/src/mfem/problem_composers/MFEMWeakFormProblemComposer.C
@@ -11,11 +11,12 @@
 
 #include "MFEMWeakFormProblemComposer.h"
 #include "EquationSystemProblemOperator.h"
+#include "MFEMProblem.h"
 
 registerMooseObject("MooseApp", MFEMWeakFormProblemComposer);
 
 MFEMWeakFormProblemComposer::MFEMWeakFormProblemComposer(const InputParameters & parameters)
-  : MFEMProblemComposer(parameters)
+  : MFEMWeakFormProblemComposerBase(parameters)
 {
 }
